requirements.txt
blinker==1.8.2
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
googletrans==3.1.0a0
gTTS==2.3.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-vlc==3.0.20123
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
urllib3==2.2.2
Werkzeug==3.0.4
zipp==3.20.0
